pyproject.toml
[project]
name = "beatool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "colorama>=0.4.6",
    "libusb>=1.0.29",
    "pyfiglet>=1.0.3",
    "pyusb>=1.3.1",
    "tqdm>=4.67.1",
    "urllib3>=2.5.0",
    "yaspin>=3.1.0",
]
